Split raw socket logic from iptables kernel module, add udp over ipv6 support
